editoast: move rolling stock common to schemas
